some changes before review
