--- app/src/main/java/ru/practicum/android/diploma/ui/countries/viewmodel/CountriesState.kt
@@ -1,16 +1,0 @@
-package ru.practicum.android.diploma.ui.countries.viewmodel
-
-import ru.practicum.android.diploma.domain.models.Country
-
-sealed interface CountriesState {
-
-    object Loading : CountriesState
-
-    data class Content(
-        val data: List<Country>
-    ) : CountriesState
-
-    data class Error(
-        val message: String
-    ) : CountriesState
-}
