+++ app/src/main/java/ru/practicum/android/diploma/ui/filter/workplace/regions/viewmodel/RegionsState.kt
@@ -1,0 +1,10 @@
+package ru.practicum.android.diploma.ui.filter.workplace.regions.viewmodel
+
+import ru.practicum.android.diploma.domain.models.Region
+
+sealed interface RegionsState {
+
+    object Loading : RegionsState
+    data class Error(val message: String) : RegionsState
+    data class Content(val data: List<Region>) : RegionsState
+}
