+++ app/src/main/res/layout/fragment_filter_industry.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white_black"
-    tools:context=".ui.industry.fragment.FilterIndustryFragment">
+    tools:context=".ui.filter.industry.fragment.FilterIndustryFragment">
 
     <LinearLayout
         android:id="@+id/industryTopNavigationBar"
