+++ app/src/main/res/layout/fragment_filter_place_of_work.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:background="@color/white_black"
-    tools:context=".ui.workplace.fragment.FilterWorkPlaceFragment">
+    tools:context=".ui.filter.workplace.fragment.FilterWorkPlaceFragment">
 
     <LinearLayout
         android:id="@+id/ll_main_text"
