+++ app/src/main/res/layout/fragment_filter_region.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white_black"
-    tools:context=".ui.regions.fragment.RegionsWorkPlaceFragment">
+    tools:context=".ui.filter.workplace.regions.fragment.RegionsWorkPlaceFragment">
 
     <LinearLayout
         android:id="@+id/top_bar"
